// legacy/krakenegg-production/versions/v1.0.0/src/types.ts
export interface FileInfo {
  id: string;
  name: string;
  path: string;
  size: number;
  modified: Date;
  created: Date;
  accessed?: Date;
  isDirectory: boolean;
  isHidden: boolean;
  permissions: FilePermissions;
  extension?: string;
  type: FileType;
  icon?: string;
}

export interface FilePermissions {
  readable: boolean;
  writable: boolean;
  executable: boolean;
  owner: string;
  group: string;
  mode: string;
}

export enum FileType {
  Directory = 'directory',
  File = 'file',
  Archive = 'archive',
  Image = 'image',
  Video = 'video',
  Audio = 'audio',
  Document = 'document',
  Code = 'code',
  Executable = 'executable',
  Link = 'link'
}